refactor: rename apply-hydrate to signals
