+++ packages/translator/src/visitors/scriptlet.ts
@@ -1,5 +1,5 @@
 import type { types as t } from "@marko/compiler";
-import { addStatement } from "../util/apply-hydrate";
+import { addStatement } from "../util/signals";
 import { isOutputHTML } from "../util/marko-config";
 import type { ReferenceGroup } from "../util/references";
 import { getSectionId } from "../util/sections";
